requirements.txt
# Core
python-dotenv>=1.0.0

# Data
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
pyarrow>=14.0.0

# API / HTTP
requests>=2.31.0

# Database
duckdb>=0.9.0

# Entity Resolution
thefuzz>=0.20.0
python-Levenshtein>=0.21.0
recordlinkage>=0.15

# ML
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.1.0
shap>=0.43.0

# Frontend
streamlit>=1.28.0
plotly>=5.17.0
altair>=5.1.0

# Testing
pytest>=7.4.0

# Optional profiling
ydata-profiling>=4.6.0
